openwrt: bind to logical interface
The interface name defined in /etc/config/network is called
logic interface name in OpenWRT. Usually, it didn't present
the interface name in Linux system.

When we configure the smartdns bind to a interface, it usually
means only the addresses assgined with that interface should be
listened. We could have many applications bind to the same port.